pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "embodied-ai-radar"
version = "0.1.0"
description = "Automated research monitoring for Embodied AI / Robotics"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "claude-agent-sdk>=0.0.3",
    "pyyaml>=6.0",
    "feedparser>=6.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "rich>=13.0",
    "click>=8.1",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
radar-daily = "radar.cli:daily"
radar-weekly = "radar.cli:weekly"
radar-deep-dive = "radar.cli:deep_dive"

[tool.hatch.build.targets.wheel]
packages = ["src/radar"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
